Remove logging and custom exceptions from the repositories
